Dynamic types
Updating to support additional file types for download dynamically, and adding additional information.

+++ files.py
@@ -1,15 +1,22 @@
 # https://stackoverflow.com/questions/11131810/passing-a-variable-in-url
+from pathlib import Path
 import requests
 import os.path
 
 def downloadFile(name,theurl,file_path):
+    print(theurl)
     r=requests.get(theurl)
-    print("****Connected****")
     completeName = os.path.expanduser(file_path + "/" + name + ".mp4")
-    f=open(completeName,'wb');
-    print("Donloading ",name," ......")
-    for chunk in r.iter_content(chunk_size=255):
-        if chunk: # filter out keep-alive new chunks
-            f.write(chunk)
-    print("Done")
-    f.close()
+    downloadPath = Path(completeName)
+    if not downloadPath.is_file():
+        print("****Connected****")
+        print("Downloading: ",name)
+        f=open(completeName,'wb');
+        print("Donloading ",name," ......")
+        for chunk in r.iter_content(chunk_size=255):
+            if chunk: # filter out keep-alive new chunks
+                f.write(chunk)
+        print("Done")
+        f.close()
+    else:
+        print("Skipped file: " + name + " because it already existed.")
